constellation-python/examples/web-demo/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
anthropic>=0.8.0
jinja2>=3.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
aiofiles>=23.0.0